requirements.txt
aiohttp>=3.9.0
pandas>=2.1.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-binance>=1.0.19
aiosqlite>=0.19.0
numpy>=1.26.0
sqlalchemy>=2.0.0

# New dependencies for production refactoring
python-json-logger>=2.0.0
statsd>=4.0.0
cryptography>=42.0.0
sortedcontainers>=2.4.0
